Got detections lines to render.

+++ src/wall.cpp
@@ -10,32 +10,32 @@
 #include <glm/gtx/string_cast.hpp>
 
 Wall::Wall(glm::vec3 pointA, glm::vec3 pointB) {
-    float Ax = pointA.x;
-    float Ay = pointA.y;
-    float Bx = pointB.x;
-    float By = pointB.y;
+    double Ax = pointA.x;
+    double Ay = pointA.y;
+    double Bx = pointB.x;
+    double By = pointB.y;
     
-    float midX = (Ax + Bx) / 2;
-    float midY = (Ay + By) / 2;
+    double midX = (Ax + Bx) / 2;
+    double midY = (Ay + By) / 2;
     
-    center = glm::vec3(midX,midY,0.0f);
+    center = glm::vec3(midX,midY,0.0);
     
     float rotation = atan2(By - Ay, Bx - Ax);
     
-    float distance = sqrt(pow(Bx - Ax, 2) + pow(By - Ay, 2));
+    double distance = sqrt(pow(Bx - Ax, 2) + pow(By - Ay, 2));
     
     length = distance;
     
-    glm::mat4 scaleMatrix = glm::scale(glm::mat4(1.0f), glm::vec3(length/2,1.0f,1.0f));
-    glm::mat4 translateMatrix = glm::translate(glm::mat4(1.0f), center);
-    glm::mat4 rotateMatrix = glm::rotate(glm::mat4(1.0f), rotation, glm::vec3(0.0f,0.0f,1.0f));
+    glm::mat4 scaleMatrix = glm::scale(glm::mat4(1.0), glm::vec3(length/2,1.0,1.0));
+    glm::mat4 translateMatrix = glm::translate(glm::mat4(1.0), center);
+    glm::mat4 rotateMatrix = glm::rotate(glm::mat4(1.0), rotation, glm::vec3(0.0,0.0,1.0));
     modelMatrix = translateMatrix * rotateMatrix * scaleMatrix;
     angle = rotation;
     
     printf("PointA: %s\nPointB: %s\n midX: %f midY: %f rotation: %f\n",glm::to_string(pointA).c_str(),glm::to_string(pointB).c_str(),midX,midY,rotation);
 }
 
-float Wall::getLength() {
+double Wall::getLength() {
     return length;
 }
 
@@ -47,6 +47,6 @@ glm::vec3 Wall::getCenter() {
     return center;
 }
 
-float Wall::getAngle() {
+double Wall::getAngle() {
     return angle;
 }
